Extract WaylandExecutor
