+++ src/server/frontend/wayland/wayland_connector.cpp
@@ -18,6 +18,9 @@
 
 #include "wayland_connector.h"
 
+#include "wayland_executor.h"
+#include "wlshmbuffer.h"
+
 #include "core_generated_interfaces.h"
 #include "xdg_shell_generated_interfaces.h"
 
@@ -43,8 +46,6 @@
 #include "mir/frontend/buffer_stream_id.h"
 #include "mir/frontend/display_changer.h"
 
-#include "mir/executor.h"
-
 #include "mir/client/event.h"
 
 #include "mir/input/device.h"
@@ -58,7 +59,6 @@
 #include <unordered_map>
 #include <boost/throw_exception.hpp>
 
-#include <future>
 #include <functional>
 #include <type_traits>
 
@@ -67,17 +67,12 @@
 #include <algorithm>
 #include <mir/log.h>
 #include <cstring>
-#include <deque>
-#include MIR_SERVER_GL_H
-#include MIR_SERVER_GLEXT_H
 
 #include "mir/fd.h"
-#include "../../../platforms/common/server/shm_buffer.h"
 
 #include <sys/stat.h>
 #include <sys/socket.h>
 #include "mir/anonymous_shm_file.h"
-#include "wlshmbuffer.h"
 
 namespace mf = mir::frontend;
 namespace mg = mir::graphics;
@@ -2563,142 +2558,6 @@ void cleanup_display(wl_display *display)
     wl_display_flush_clients(display);
     wl_display_destroy(display);
 }
-
-class WaylandExecutor : public mir::Executor
-{
-public:
-    void spawn (std::function<void ()>&& work) override
-    {
-        {
-            std::lock_guard<std::recursive_mutex> lock{mutex};
-            workqueue.emplace_back(std::move(work));
-        }
-        if (auto err = eventfd_write(notify_fd, 1))
-        {
-            BOOST_THROW_EXCEPTION((std::system_error{err, std::system_category(), "eventfd_write failed to notify event loop"}));
-        }
-    }
-
-    /**
-     * Get an Executor which dispatches onto a wl_event_loop
-     *
-     * \note    The executor may outlive the wl_event_loop, but no tasks will be dispatched
-     *          after the wl_event_loop is destroyed.
-     *
-     * \param [in]  loop    The event loop to dispatch on
-     * \return              An Executor that queues onto the wl_event_loop
-     */
-    static std::shared_ptr<mir::Executor> executor_for_event_loop(wl_event_loop* loop)
-    {
-        if (auto notifier = wl_event_loop_get_destroy_listener(loop, &on_display_destruction))
-        {
-            DestructionShim* shim;
-            shim = wl_container_of(notifier, shim, destruction_listener);
-
-            return shim->executor;
-        }
-        else
-        {
-            auto const executor = std::shared_ptr<WaylandExecutor>{new WaylandExecutor{loop}};
-            auto shim = std::make_unique<DestructionShim>(executor);
-
-            shim->destruction_listener.notify = &on_display_destruction;
-            wl_event_loop_add_destroy_listener(loop, &(shim.release())->destruction_listener);
-
-            return executor;
-        }
-    }
-
-private:
-    WaylandExecutor(wl_event_loop* loop)
-        : notify_fd{eventfd(0, EFD_CLOEXEC | EFD_SEMAPHORE | EFD_NONBLOCK)},
-          notify_source{wl_event_loop_add_fd(loop, notify_fd, WL_EVENT_READABLE, &on_notify, this)}
-    {
-        if (notify_fd == mir::Fd::invalid)
-        {
-            BOOST_THROW_EXCEPTION((std::system_error{
-                errno,
-                std::system_category(),
-                "Failed to create IPC pause notification eventfd"}));
-        }
-    }
-
-    std::function<void()> get_work()
-    {
-        std::lock_guard<std::recursive_mutex> lock{mutex};
-        if (!workqueue.empty())
-        {
-            auto const work = std::move(workqueue.front());
-            workqueue.pop_front();
-            return work;
-        }
-        return {};
-    }
-
-    static int on_notify(int fd, uint32_t, void* data)
-    {
-        auto executor = static_cast<WaylandExecutor*>(data);
-
-        eventfd_t unused;
-        if (auto err = eventfd_read(fd, &unused))
-        {
-            mir::log_error(
-                "eventfd_read failed to consume wakeup notification: %s (%i)",
-                strerror(err),
-                err);
-        }
-
-        while (auto work = executor->get_work())
-        {
-            try
-            {
-                work();
-            }
-            catch(...)
-            {
-                mir::log(
-                    mir::logging::Severity::critical,
-                    MIR_LOG_COMPONENT,
-                    std::current_exception(),
-                    "Exception processing Wayland event loop work item");
-            }
-        }
-
-        return 0;
-    }
-
-    static void on_display_destruction(wl_listener* listener, void*)
-    {
-        DestructionShim* shim;
-        shim = wl_container_of(listener, shim, destruction_listener);
-
-        {
-            std::lock_guard<std::recursive_mutex> lock{shim->executor->mutex};
-            wl_event_source_remove(shim->executor->notify_source);
-        }
-        delete shim;
-    }
-
-    std::recursive_mutex mutex;
-    mir::Fd const notify_fd;
-    std::deque<std::function<void()>> workqueue;
-
-    wl_event_source* const notify_source;
-
-    struct DestructionShim
-    {
-        explicit DestructionShim(std::shared_ptr<WaylandExecutor> const& executor)
-            : executor{executor}
-        {
-        }
-
-        std::shared_ptr<WaylandExecutor> const executor;
-        wl_listener destruction_listener;
-    };
-    static_assert(
-        std::is_standard_layout<DestructionShim>::value,
-        "DestructionShim must be Standard Layout for wl_container_of to be defined behaviour");
-};
 }
 
 mf::WaylandConnector::WaylandConnector(
@@ -2735,7 +2594,7 @@ mf::WaylandConnector::WaylandConnector(
      * So far I've only found ones which expect wl_compositor before anything else,
      * so stick that first.
      */
-    auto const executor = WaylandExecutor::executor_for_event_loop(wl_display_get_event_loop(display.get()));
+    auto const executor = executor_for_event_loop(wl_display_get_event_loop(display.get()));
 
     compositor_global = std::make_unique<mf::WlCompositor>(
         display.get(),
@@ -2841,7 +2700,7 @@ int mf::WaylandConnector::client_socket_fd() const
     else
     {
         // TODO: Wait on the result of wl_client_create so we can throw an exception on failure.
-        WaylandExecutor::executor_for_event_loop(wl_display_get_event_loop(display.get()))
+        executor_for_event_loop(wl_display_get_event_loop(display.get()))
             ->spawn(
                 [socket = socket_fd[server], display = display.get()]()
                 {
@@ -2869,7 +2728,7 @@ int mf::WaylandConnector::client_socket_fd(
 
 void mf::WaylandConnector::run_on_wayland_display(std::function<void(wl_display*)> const& functor)
 {
-    auto executor = WaylandExecutor::executor_for_event_loop(wl_display_get_event_loop(display.get()));
+    auto executor = executor_for_event_loop(wl_display_get_event_loop(display.get()));
 
     executor->spawn([display_ref = display.get(), functor]() { functor(display_ref); });
 }
