+++ src/server/frontend/wayland/wayland_executor.cpp
@@ -1,0 +1,167 @@
+/*
+ * Copyright © 2015-2018 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Authored by: Christopher James Halse Rogers <christopher.halse.rogers@canonical.com>
+ */
+
+#include "wayland_executor.h"
+
+#include "mir/fd.h"
+#include "mir/log.h"
+
+#include <sys/eventfd.h>
+
+#include <boost/throw_exception.hpp>
+
+#include <cstring>
+#include <deque>
+#include <functional>
+#include <mutex>
+#include <system_error>
+
+namespace
+{
+class WaylandExecutor : public mir::Executor
+{
+public:
+    void spawn (std::function<void ()>&& work) override
+    {
+        {
+            std::lock_guard<std::recursive_mutex> lock{mutex};
+            workqueue.emplace_back(std::move(work));
+        }
+        if (auto err = eventfd_write(notify_fd, 1))
+        {
+            BOOST_THROW_EXCEPTION((std::system_error{err, std::system_category(), "eventfd_write failed to notify event loop"}));
+        }
+    }
+
+    static std::shared_ptr<mir::Executor> executor_for_event_loop(wl_event_loop* loop)
+    {
+        if (auto notifier = wl_event_loop_get_destroy_listener(loop, &on_display_destruction))
+        {
+            DestructionShim* shim;
+            shim = wl_container_of(notifier, shim, destruction_listener);
+
+            return shim->executor;
+        }
+        else
+        {
+            auto const executor = std::shared_ptr<WaylandExecutor>{new WaylandExecutor{loop}};
+            auto shim = std::make_unique<DestructionShim>(executor);
+
+            shim->destruction_listener.notify = &on_display_destruction;
+            wl_event_loop_add_destroy_listener(loop, &(shim.release())->destruction_listener);
+
+            return executor;
+        }
+    }
+
+private:
+    WaylandExecutor(wl_event_loop* loop)
+        : notify_fd{eventfd(0, EFD_CLOEXEC | EFD_SEMAPHORE | EFD_NONBLOCK)},
+        notify_source{wl_event_loop_add_fd(loop, notify_fd, WL_EVENT_READABLE, &on_notify, this)}
+    {
+        if (notify_fd == mir::Fd::invalid)
+        {
+            BOOST_THROW_EXCEPTION((std::system_error{
+                errno,
+                std::system_category(),
+                "Failed to create IPC pause notification eventfd"}));
+        }
+    }
+
+    std::function<void()> get_work()
+    {
+        std::lock_guard<std::recursive_mutex> lock{mutex};
+        if (!workqueue.empty())
+        {
+            auto const work = std::move(workqueue.front());
+            workqueue.pop_front();
+            return work;
+        }
+        return {};
+    }
+
+    static int on_notify(int fd, uint32_t, void* data)
+    {
+        auto executor = static_cast<WaylandExecutor*>(data);
+
+        eventfd_t unused;
+        if (auto err = eventfd_read(fd, &unused))
+        {
+            mir::log_error(
+                "eventfd_read failed to consume wakeup notification: %s (%i)",
+                strerror(err),
+                err);
+        }
+
+        while (auto work = executor->get_work())
+        {
+            try
+            {
+                work();
+            }
+            catch(...)
+            {
+                mir::log(
+                    mir::logging::Severity::critical,
+                    MIR_LOG_COMPONENT,
+                    std::current_exception(),
+                    "Exception processing Wayland event loop work item");
+            }
+        }
+
+        return 0;
+    }
+
+    static void on_display_destruction(wl_listener* listener, void*)
+    {
+        DestructionShim* shim;
+        shim = wl_container_of(listener, shim, destruction_listener);
+
+        {
+            std::lock_guard<std::recursive_mutex> lock{shim->executor->mutex};
+            wl_event_source_remove(shim->executor->notify_source);
+        }
+        delete shim;
+    }
+
+    std::recursive_mutex mutex;
+    mir::Fd const notify_fd;
+    std::deque<std::function<void()>> workqueue;
+
+    wl_event_source* const notify_source;
+
+    struct DestructionShim
+    {
+        explicit DestructionShim(std::shared_ptr<WaylandExecutor> const& executor)
+            : executor{executor}
+        {
+        }
+
+        std::shared_ptr<WaylandExecutor> const executor;
+        wl_listener destruction_listener;
+    };
+    static_assert(
+        std::is_standard_layout<DestructionShim>::value,
+        "DestructionShim must be Standard Layout for wl_container_of to be defined behaviour");
+};
+}
+
+std::shared_ptr<mir::Executor> mir::frontend::executor_for_event_loop(wl_event_loop* loop)
+{
+    return WaylandExecutor::executor_for_event_loop(loop);
+}
