+++ src/server/frontend/wayland/wlshmbuffer.cpp
@@ -24,8 +24,8 @@
 
 #include <wayland-server-protocol.h>
 
-#include <GLES2/gl2.h>
-#include <GLES2/gl2ext.h>
+#include MIR_SERVER_GL_H
+#include MIR_SERVER_GLEXT_H
 
 #include <boost/throw_exception.hpp>
